+++ src/scenes/UIScene.js
@@ -8,6 +8,7 @@ import SystemLog from '../ui/SystemLog.js';
 import ZoneNav from '../ui/ZoneNav.js';
 import BossChallenge from '../ui/BossChallenge.js';
 import DrinkButton from '../ui/DrinkButton.js';
+import SmashButton from '../ui/SmashButton.js';
 import InventoryPanel from '../ui/InventoryPanel.js';
 import UpgradePanel from '../ui/UpgradePanel.js';
 import PrestigePanel from '../ui/PrestigePanel.js';
@@ -52,6 +53,7 @@ export default class UIScene extends Phaser.Scene {
     this.zoneNav = new ZoneNav(this);
     this.bossChallenge = new BossChallenge(this);
     this.drinkButton = new DrinkButton(this);
+    this.smashButton = new SmashButton(this);
     this.inventoryPanel = new InventoryPanel(this);
     this.upgradePanel = new UpgradePanel(this);
     this.settingsPanel = new SettingsPanel(this);
@@ -131,12 +133,14 @@ export default class UIScene extends Phaser.Scene {
       this.zoneNav.hide();
       this.bossChallenge.hide();
       this.drinkButton.hide();
+      this.smashButton.hide();
       this._mapOpen = true;
     } else {
       overworldScene.scene.sleep();
       this.zoneNav.show();
       this.bossChallenge.show();
       this.drinkButton.show();
+      this.smashButton.show();
       this._mapOpen = false;
     }
   }
@@ -148,6 +152,7 @@ export default class UIScene extends Phaser.Scene {
     if (this.zoneNav) { this.zoneNav.destroy(); this.zoneNav = null; }
     if (this.bossChallenge) { this.bossChallenge.destroy(); this.bossChallenge = null; }
     if (this.drinkButton) { this.drinkButton.destroy(); this.drinkButton = null; }
+    if (this.smashButton) { this.smashButton.destroy(); this.smashButton = null; }
     for (const modal of this._modals || []) {
       if (modal) modal.destroy();
     }